% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copc4R-package.R
\docType{package}
\name{copc4R-package}
\alias{copc4R}
\alias{copc4R-package}
\title{copc4R: Read COPC Point Clouds from Local Files and HTTP URLs}
\description{
Reads and writes Cloud Optimized Point Cloud (COPC, .copc.laz) files from local disk or HTTP(S) endpoints using range reads. Returns point data as 'data.table' with 'rlas'-compatible header conventions. Optionally converts to 'lidR' LAS objects when 'lidR' is installed. Uses 'LASzip' for LAZ decompression. Supports spatial bounding box, polygon area-of-interest (AOI), corridor, Z-range, and attribute filters via the COPC octree hierarchy. Provides multi-resolution level-of-detail (LOD) sampling, deterministic tiling, streaming iterators, 'lidR' catalog integration, STAC endpoint discovery (including the USGS 3DEP collection via Microsoft Planetary Computer), Virtual Point Cloud (VPC) read/write for QGIS/PDAL interoperability, and octree-based adaptive chunking for balanced work-unit processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bi0m3trics/copc4R}
  \item Report bugs at \url{https://github.com/bi0m3trics/copc4R/issues}
}

}
\author{
\strong{Maintainer}: Andrew Sánchez Meador \email{andrew.meador@nau.edu}

Other contributors:
\itemize{
  \item Martin Isenburg (LASzip library) [copyright holder]
  \item Jean-Romain Roussel (rlas conventions and LASzip R modifications) [copyright holder]
  \item Andrew Bell (COPC specification (Hobu, Inc.)) [copyright holder]
  \item Howard Butler (COPC specification (Hobu, Inc.)) [copyright holder]
  \item Connor Manning (COPC specification (Hobu, Inc.)) [copyright holder]
}

}
\keyword{internal}
