% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{copc_bounds}
\alias{copc_bounds}
\title{Get COPC file spatial bounds}
\usage{
copc_bounds(path_or_url, as_sf = FALSE)
}
\arguments{
\item{path_or_url}{Character. Path or URL to a \code{.copc.laz} file.}

\item{as_sf}{Logical.  If \code{TRUE} and \pkg{sf} is installed,
return an \code{sf} bounding-box polygon with CRS.  Default \code{FALSE}.}
}
\value{
If \code{as_sf = FALSE}, a named numeric vector
\code{c(xmin, ymin, zmin, xmax, ymax, zmax)}.
If \code{as_sf = TRUE}, an \code{sf::st_bbox()} object.
}
\description{
Returns the 3-D bounding box from the LAS header.  No points are read.
}
\examples{
f <- system.file("extdata", "NoAZCampus_SWFSC_Bldg82_2mVoxel.copc.laz",
                 package = "copc4R")
copc_bounds(f)

\donttest{
if (requireNamespace("sf", quietly = TRUE)) {
  copc_bounds(f, as_sf = TRUE)
}
}

}
