% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{copc_cache_clear}
\alias{copc_cache_clear}
\title{Clear the COPC cache}
\usage{
copc_cache_clear(mem = TRUE, disk = FALSE)
}
\arguments{
\item{mem}{Logical.  Clear the in-memory LRU cache? Default \code{TRUE}.}

\item{disk}{Logical.  Clear the on-disk persistent cache? Default \code{FALSE}.}
}
\value{
Invisibly returns the cache stats \emph{before} clearing.
}
\description{
Clear the COPC cache
}
\examples{
copc_cache_clear(mem = TRUE, disk = FALSE)

}
