% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{copc_cache_stats}
\alias{copc_cache_stats}
\title{Report cache statistics}
\usage{
copc_cache_stats()
}
\value{
A named list with \code{mem_entries}, \code{mem_mb},
\code{disk_entries}, \code{disk_mb}, \code{disk_dir}.
}
\description{
Report cache statistics
}
\examples{
copc_cache_stats()

}
