% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copc.4R-package.R, R/copc4R-package.R
\docType{package}
\name{copc4R-package}
\alias{copc4R-package}
\title{copc4R: Read COPC Point Clouds from Local Files and HTTP URLs}
\description{
Reads Cloud Optimized Point Cloud (COPC, .copc.laz) files from local disk or HTTP endpoints using range reads. Returns point data as data.table with rlas-compatible header conventions. Optionally converts to lidR::LAS objects when lidR is installed. Uses LASzip for LAZ decompression. Supports spatial bounding box and Z-range queries via the COPC octree hierarchy.

Reads Cloud Optimized Point Cloud (COPC, .copc.laz) files from local disk or HTTP endpoints using range reads. Returns point data as data.table with rlas-compatible header conventions. Optionally converts to lidR::LAS objects when lidR is installed. Uses LASzip for LAZ decompression. Supports spatial bounding box and Z-range queries via the COPC octree hierarchy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bi0m3trics/copc4R}
  \item Report bugs at \url{https://github.com/bi0m3trics/copc4R/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/bi0m3trics/copc4R}
  \item Report bugs at \url{https://github.com/bi0m3trics/copc4R/issues}
}

}
\author{
\strong{Maintainer}: Andrew Sánchez Meador \email{andrew.meador@nau.edu}

Other contributors:
\itemize{
  \item Martin Isenburg (LASzip library) [copyright holder]
  \item Jean-Romain Roussel (rlas conventions and LASzip R modifications) [copyright holder]
  \item Andrew Bell (COPC specification (Hobu, Inc.)) [copyright holder]
  \item Howard Butler (COPC specification (Hobu, Inc.)) [copyright holder]
  \item Connor Manning (COPC specification (Hobu, Inc.)) [copyright holder]
}

}
\keyword{internal}
